git tracking issues resolved. solution up to date.
