+++ ConsoleApp1/Models/EmployeeData.cs
@@ -1,0 +1,9 @@
+﻿namespace PayCal.Models
+{
+    public abstract class EmployeeData
+    {
+        public int EmployeeID { get; set; }
+        public string FName { get; set; }
+        public string LName { get; set; }
+    }
+}
