+++ ConsoleApp1/Services/ICalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace PayCal.Services
+{
+    public interface ICalculator
+    {
+        (double, double) CalculateEmployeePay(int employeeID);
+    }
+}
