+++ ConsoleApp2/TempEmployeeDataShould.cs
@@ -1,15 +1,16 @@
 ﻿using NUnit.Framework;
+using PayCal.Models;
 
 namespace PayCal___Tests
 {
     [TestFixture]
-    public class PermEmployeeDataShould
+    public class TempEmployeeDataShould
     {
         [Test]
         public void Check_FName_is_null()
         {
             // Arrange
-            var sut = new PermEmployeeData();
+            var sut = new TempEmployeeData();
 
             // Act + Assert
             Assert.Null(sut.FName);
@@ -19,30 +20,30 @@ namespace PayCal___Tests
         public void Check_LName_is_null()
         {
             // Arrange
-            var sut = new PermEmployeeData();
+            var sut = new TempEmployeeData();
 
             // Act + Assert
             Assert.Null(sut.LName);
         }
 
         [Test]
-        public void Check_Salaryint_is_null()
+        public void Check_DayRateint_is_null()
         {
             // Arrange
-            var sut = new PermEmployeeData();
+            var sut = new TempEmployeeData();
 
             // Act + Assert
-            Assert.Null(sut.Salaryint);
+            Assert.Null(sut.DayRateint);
         }
 
         [Test]
-        public void Check_Bonusint_is_null()
+        public void Check_WeeksWorkedint_is_null()
         {
             // Arrange
-            var sut = new PermEmployeeData();
+            var sut = new TempEmployeeData();
 
             // Act + Assert
-            Assert.Null(sut.Bonusint);
+            Assert.Null(sut.WeeksWorkedint);
         }
 
         [Test]
