+++ PayCal API.sln
@@ -4,16 +4,12 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PayCal API", "PayCal API\PayCal API.csproj", "{CE04295F-14CC-4100-9133-020BC4EE6C5A}"
-	ProjectSection(ProjectDependencies) = postProject
-		{8441060E-AD54-4490-A437-B1F655265EFB} = {8441060E-AD54-4490-A437-B1F655265EFB}
-		{A1C826FE-D417-4691-9B04-EAB682BD761D} = {A1C826FE-D417-4691-9B04-EAB682BD761D}
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PayCal", "..\PayCal\PayCal\PayCal.csproj", "{A995A623-45D0-48AA-9CA0-BF7380594DC3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PayCal - Tests", "..\PayCal\PayCal - Tests\PayCal - Tests.csproj", "{8441060E-AD54-4490-A437-B1F655265EFB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PayCal Console", "PayCal Console\PayCal Console.csproj", "{A1C826FE-D417-4691-9B04-EAB682BD761D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PayCal", "ConsoleApp1\PayCal.csproj", "{46F7888A-80F6-417B-A647-BE3603FC51F8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PayCal - Tests", "ConsoleApp2\PayCal - Tests.csproj", "{F798986A-7999-49E4-A0CF-7D9423538113}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -25,18 +21,18 @@ Global
 		{CE04295F-14CC-4100-9133-020BC4EE6C5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CE04295F-14CC-4100-9133-020BC4EE6C5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CE04295F-14CC-4100-9133-020BC4EE6C5A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A995A623-45D0-48AA-9CA0-BF7380594DC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A995A623-45D0-48AA-9CA0-BF7380594DC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A995A623-45D0-48AA-9CA0-BF7380594DC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A995A623-45D0-48AA-9CA0-BF7380594DC3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8441060E-AD54-4490-A437-B1F655265EFB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8441060E-AD54-4490-A437-B1F655265EFB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8441060E-AD54-4490-A437-B1F655265EFB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8441060E-AD54-4490-A437-B1F655265EFB}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A1C826FE-D417-4691-9B04-EAB682BD761D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A1C826FE-D417-4691-9B04-EAB682BD761D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A1C826FE-D417-4691-9B04-EAB682BD761D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{A1C826FE-D417-4691-9B04-EAB682BD761D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{46F7888A-80F6-417B-A647-BE3603FC51F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{46F7888A-80F6-417B-A647-BE3603FC51F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{46F7888A-80F6-417B-A647-BE3603FC51F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{46F7888A-80F6-417B-A647-BE3603FC51F8}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F798986A-7999-49E4-A0CF-7D9423538113}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F798986A-7999-49E4-A0CF-7D9423538113}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F798986A-7999-49E4-A0CF-7D9423538113}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F798986A-7999-49E4-A0CF-7D9423538113}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
