+++ PayCal API/PayCal API.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
@@ -17,7 +17,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\PayCal\PayCal\PayCal.csproj" />
+    <ProjectReference Include="..\ConsoleApp1\PayCal.csproj" />
   </ItemGroup>
 
 </Project>
