+++ PayCal Console/PayCal Console.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
@@ -9,7 +9,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\PayCal\PayCal\PayCal.csproj" />
+    <ProjectReference Include="..\ConsoleApp1\PayCal.csproj" />
   </ItemGroup>
 
 </Project>
