--- PayCal/PayCal.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31702.278
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PayCal", "PayCal\PayCal.csproj", "{185CE7A5-E032-4260-BA8C-1545FB3045ED}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PayCal - Tests", "PayCal - Tests\PayCal - Tests.csproj", "{79DF96E8-894D-4160-AD07-124C360C20F7}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{185CE7A5-E032-4260-BA8C-1545FB3045ED}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{185CE7A5-E032-4260-BA8C-1545FB3045ED}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{185CE7A5-E032-4260-BA8C-1545FB3045ED}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{185CE7A5-E032-4260-BA8C-1545FB3045ED}.Release|Any CPU.Build.0 = Release|Any CPU
-		{79DF96E8-894D-4160-AD07-124C360C20F7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{79DF96E8-894D-4160-AD07-124C360C20F7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{79DF96E8-894D-4160-AD07-124C360C20F7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{79DF96E8-894D-4160-AD07-124C360C20F7}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {26606242-8B98-4090-95FD-BAFDCF307EA4}
-	EndGlobalSection
-EndGlobal
