--- PayCal/PayCal/ICalculator.cs
@@ -1,7 +1,0 @@
-﻿namespace PayCal.Services
-{
-    public interface ICalculator
-    {
-        (double, double) CalculateEmployeePay(int employeeID);
-    }
-}
